Code refactoring to split all configuration logic into a crate
This was needed to avoid circular dependencies after the project's
reorganisation into a workspace.

[GH-25]
