+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["bot-core", "clientlib"]
+members = ["bot-core", "clientlib", "configuration"]
 
 [workspace.package]
 edition = "2024"
