+++ configuration/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "configuration"
+edition.workspace = true
+version.workspace = true
+license.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+config = "0.15.11"
+secrecy = "0.10.3"
+sqlx = { version = "0.8", features = [ "runtime-tokio", "postgres", "mysql", "macros", "uuid", "chrono" ] }
+serde = { version = "1.0.217", features = ["serde_derive"] }
+serde_derive = "1.0"
+
+[lints]
+workspace = true
